Phase-2/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.16
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0